Move validation of Configuration to where it's set
Previously, the Configuration of CommandLineArguments was validated in the TryParse() and DisplayHelp() methods. However, that means that we validate the configuration every time we call one of those methods.

Moreover, stacktraces/errors can be confusing for the users. They get errors when they use the config, not when they set it. But, that can be in a totally different part of the application. Now they get errors when they set an invalid config.