Remove duplicate property finding in ArgumentPropertiesHelper.FillProperty<T>() (fixes #33)
Previously, for each property we wanted to set we looped through all properties in T to find the correct one. However, at the call site of FillProperty<T> we most likely already know which property we're dealing with so we can operate directly on that PropertyInfo.

This is a significant performance improvement and cumulative too, because we searched the properties for each property and now we only do that once.