Simplified code of PrefixedKeyParser.TryFindRawArgumentValue() (fixes #64)
Previously, the code of PrefixedKeyParser.TryFindRawArgumentValue() was incredibly complex with lots of nested conditions. And in the worst case, every array element after the first one was evaluated twice.

This change simplifies the code greatly, reduces branching a lot and in general is around 1.5-2 times faster than the old code with the same memory usage.